Anonymized Research - Front End
In the Lisbon HackEvent as part of the Anonymized research team.
To allow an eventual "researcher" profile to have access to an anonymized
version of the data the plan is to have second database in the application which
contains the data that the researcher will have access to.

From a front-end point of view this requires the selection of the database to be dynamic
and binded to the user role. This commit makes that part of the process.

In the database service, the "anon" database is initalized and stored in a "anonDB" property
of the service (the name of the database in hardcoded to "anon" (both remote and local). The setup function
now creates two instances of PouchDB and the "mainDB" property is now a computed value that
depends on the userRole.

Missing is still the part where the "researcher" role is created and is able to perform anything useful.
Couldn't find a way to create a new role.

Also updates the service workers to now replicate the remote database to the local database (but not the other way around)
The name of the database in the service workers is still hardcoded to 'anon'.

Also missing is updating any initialization scripts and README changes so that users know that an 'anon' database
must exist both as a local pouchdb database and a remote couchdb database.